// lib/api/generated/types.gen.ts
export type StrategyResponse = {
    /**
     * Id
     */
    id: number;
    /**
     * Name
     */
    name: string;
    /**
     * Strategy Json
     */
    strategy_json: {
        [key: string]: unknown;
    };
};
// ...
export type StrategySignalEvaluationResult = {
    /**
     * Strategy Id
     */
    strategy_id: number;
    /**
     * Strategy Name
     */
    strategy_name: string;
    /**
     * Asset
     */
    asset: string;
    /**
     * Interval
     */
    interval: string;
    /**
     * Timestamp
     */
    timestamp?: string | null;
    /**
     * Entry Strength
     */
    entry_strength?: number | null;
    /**
     * Exit Strength
     */
    exit_strength?: number | null;
    /**
     * Entry Threshold
     */
    entry_threshold?: number;
    /**
     * Exit Threshold
     */
    exit_threshold?: number;
    /**
     * Signals
     */
    signals?: Array<string>;
};